requirements/base.txt
django
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
django-allauth==0.51.0
django-rest-auth
social-auth-app-django==5.0.0
social-auth-core==4.3.0
djangorestframework-simplejwt==5.2.2
dj-rest-auth==2.2.5
Pillow
django-ckeditor==6.5.1
django-modeltranslation==0.18.9
django-rosetta==0.9.9
django-resized==1.0.2
django-auditlog==2.2.2
django-recaptcha==3.0.0
django-debug-toolbar==4.0.0
sentry-sdk==1.22.2
redis==4.5.1
mutagen==1.45.1
celery==5.2.7
